Add support for one-way FPGA advertisement pin
Apollo firmware will keep the USB switch handed over to the FPGA as
long as the gateware keeps advertising its usage through the FPGA_ADV
pin (FPGA_INT). When these messages stop arriving, Apollo will take
over the port. If the advertisement resumes, the port will not be
handed off to the FPGA again until a "reconfiguration" (0xc0) or a
"honor FPGA_ADV" (0xc2) vendor request arrives.

Additionally, the host tools can request a gateware to hand off the
USB port to Apollo if the necessary request is available.

- firmware: Adds HONOR_FPGA_ADV vendor request (0xc2)
- apollo_fpga: Use REQUEST_APOLLO_ADV_STOP (0xf0) to enumerate Apollo